Crate test_shutdown

Structs

ShutdownArguments ğŸ”’

Functions

check_shutdown_call ğŸ”’

get_tests ğŸ”’

main ğŸ”’

read_all ğŸ”’

Read from the fd until an EOF.

read_once ğŸ”’

A wrapper for libc::read() which uses a slice.

setup_dgram_sockets ğŸ”’

Generate a pair of connected UDP sockets.

setup_stream_sockets ğŸ”’

Generate a pair of connected TCP sockets.

test_after_close ğŸ”’

Test shutdown() after closing the socket.

test_arguments ğŸ”’

Test shutdown() with different arguments, making sure that the return value is expected.

test_conn_reset ğŸ”’

Test a certain case where we receive ECONNRESET when reading after shutdown().

test_invalid_fd ğŸ”’

Test shutdown() using an argument that cannot be a fd.

test_invalid_how ğŸ”’

Test shutdown() using a valid socket, but invalid how argument.

test_non_existent_fd ğŸ”’

Test shutdown() using an argument that could be a fd, but is not.

test_non_socket_fd ğŸ”’

Test shutdown() using a valid fd that is not a socket.

test_not_connected ğŸ”’

Test shutdown() using a non-connected socket.

test_read_after_client_shutdown ğŸ”’

Test reading from the socket after shutdown().

test_read_after_peer_shutdown ğŸ”’

Test reading from the socket after shutdown() on the peer.

test_tcp_fin_correctness ğŸ”’

Test that the FIN is not sent until the buffer is cleared after shutdown() on both sockets.

test_twice ğŸ”’

Test calling shutdown() twice with the same arguments.

test_write_after_client_shutdown ğŸ”’

Test writing to the socket after shutdown().

test_write_after_peer_shutdown ğŸ”’

Test writing to the socket after shutdown() on the peer.

write_all ğŸ”’

Write all bytes in the buffer.

write_once ğŸ”’

A wrapper for libc::write() which uses a slice.